// cypress/support/enums/leave-enum.ts
export enum LEAVE_KEY {
  ENTITLEMENTS = "Entitlements",
  ADD_ENTITLEMENTS = "Add Entitlements",
  LEAVE_TYPE = "CAN - FMLA",
  LEAVE_BALANCE = "10",
  CONFIRM = "Confirm",
  LOGOUT = "Logout",
  APPLY = "Apply",
  DATE_FROM = "2025-01-11",
  DATE_AFTER = "2025-10-11",
  APPROVE = "Approve",
  ALL_DAYS = "Full Days",
  HALF_DAY_MORNING = "Half Day - Morning",
  LEAVE_PERIOD = "2025-01-11 to 2025-10-11",
  CANCEL = "Cancel",
  CLOSE = "Close"
}